Infer: in serialization test, check qualification logic
Summary:
This commit leaves behavior unchanged, it's just tweaking
the GlobalResult serialization test to verify how the
`pyre infer` backend is handling type qualification.

The conclusion is that at the moment:
- an inferred type will be dequalified to account for imports
- but a given type will be fully-qualified

This isn't desirable, it would be nicer if the client could
count on the same qualification in all cases. But it isn't
*urgent* to fix because working with completely dequalified
types happens to work out in the LibCST transform (we just
wind up naively generating code without understanding the
actual type names).

So, for now I'm just testing it so I can refer to the behavior.
I'll probably put up a diff later fixing it. If you are
reviewing, feel free to comment with tips on where to look
for the type dequalification.

Reviewed By: dkgi

Differential Revision: D31774468

fbshipit-source-id: 7e12b4c4828c203193a99eae4fdc4cd846d298d6