Fix brittle parsing logic in pyre-upgrade
Summary:
We incorrectly assumed that a `# pyre-ignore [_]` directive
can only have comma-separated integers in it.

The problem is that pyre will recognize an empty `# pyre-ignore`,
and the actual ignore is user-controlled. There is in fact a
project where we're unable to run an upgrade right now due to
a line with `# pyre-ignore[]`.

The simplest solution is also the most robust to any other weird
thing that might slip in (unclear if other content is possible - probably
not because the error has to get through the ocaml parser): just skip any
data that isn't a valid int, including the empty string we hit on a
`pyre-ignore []`.

Reviewed By: grievejia

Differential Revision: D31809024

fbshipit-source-id: 0bb8951c85f4141f4fae10ac65189c3656a12b55