Infer: revive tests of should_analyze_define
Summary:
The quick fix in D31801754 (https://github.com/facebook/pyre-check/commit/12a786c7f2379d3d08643321b20fe6f125c5dec7), which I pushed out as fast as
possible to unbreak trunk, required disabling the existing
tests of should_analyze_define, which were depending on us
actually outputting nothing as opposed to outputting
empty results.

This diff revives the tests, which required adding a helper
function to help us see inside what happens (the actual output
is identical for skipped and not-skipped defines that have
full annotations). The helper funciton goes in a Testing module
so that it's clear that - even though we do have to declare it
in our interface - it should be clear that this isn't production
code.

Reviewed By: dkgi

Differential Revision: D31804472

fbshipit-source-id: 8b612ddca0723d66d380fcec1847ebef79b8bbc8