Infer: get rid of typing dequalification
Summary:
I've long been puzzled by the typing dequalification
logic - the LibCST code deals with qualification for us
(it isn't perfect, but for simple `<module>.<type>` names
like `typing.Dict` it works 100%).

I finally dug around in the history long enough to figure out
where this came from. It predated *all* namespace handling
and the LibCST code. So I'm pretty confident we can just
scrap it.

If this leads to any new bugs, we should handle them in
the LibCST visitor. Hacks like this need to die so we can
focus on a correct design.

Reviewed By: dkgi

Differential Revision: D31799391

fbshipit-source-id: f19fc72f4eb50b10ab8df7ac2fa19f4407d29d15