Infer: dequalify types from current module in the client.
Summary:
One class of bug we hit frequently in infer codemods right now
are cases where we try to import types from the current module,
producng a circular import.

The root cause is that in some cases the pyre backend dequalifies
types but in others it does not (in particular, ocaml `Type.t`
gets dequalified, but `Expression.t` does not, so given annotations
come back fully-qualified).

My view is that because annotation generation logic has to center
on LibCST for other reasons, it's probably prefarable to
consistently use fully-qualified names rather than dequalified names.

This commit tweaks the client by changing our annotation fixer to both handle
pathlike annotations and also dequalify imports from the current
module. I'll update the backend in the next commit; modifying the
client first should work pretty well - there is a weird edge case
involving inested classes whose name matches the module name but that should
be pretty rare.

The way it works is that
- I save the `qualifier` when splitting raw infer output
- I use that `qualifier` to dequalify types

NOTE: this is actually a pretty straightforward change. It's large in terms of lines
of code because: (a) I have to pass `qualifier` in many more places, especially tests
(b) quite a few test lines had to be tweaked either to reflect the dequalification
logic or to pass `qualifier around` (c) I had to introduce a new type
`RawInferOutputForPath` to reflect that the post-split data has an extra field.

NOTE: it's debateable whether handling dequalification in the client code s a good
idea - we can probably do even better by pushing it all the way down to LibCST.
But that is *not* a reason to reject this diff: to get it into LibCST we have to do this
first because (a) we need this change before we can remove dequalification from
the ocaml side (which is what I'll do next) and (b) it's much, much easier to move
dequalification from pyre's python code to LibCST's python code than to jump
all the way from ocaml to LibCST in one hop.

Reviewed By: grievejia

Differential Revision: D31837568

fbshipit-source-id: 6a27b395ddabec5e759a5f73f82d0fdb47e8ec5e