Trigger website build on open source PRs
Publishing of the Pysa website has been [broken for a few weeks](https://github.com/facebook/pyre-check/actions/workflows/publish_website.yml). #514 contains the version upgrade needed to fix this, but currently the website build is only triggering when the `main` branch is updated, rather than on PRs. This PR updates the `publish_website` workflow to also run on PRs, but in a dry run mode that skips the final publish step.

Test Plan:
If the action runs on this PR, I think that makes it a success!