Use bare file suffix in watchman filter
Summary:
Additional extensions are represented as `.extension` in our configuration object, but are expected to be passed to the watchman filter without the leading `.`.

We were actually failing to pick up on watchman updates for additional extensions (ie., .cconf) in configerator - stripping this prefix fixes the behavior.

Will try to get a binary pushed out to configerator after this lands.

Reviewed By: dkgi

Differential Revision: D31853453

fbshipit-source-id: 91d5d8041e44649e27a2990dc45d39da1fc6f7b5