Make parent file / directory tests robust to existing files
Summary:
I created files `/tmp/a` and `/tmp/b` when working on some ocaml
tests yesterday, because I wanted to diff some json output. This
broke my unit tests!

It turns out that the existing `find_parent_directory_containing_*`
logic is not possible to test robustly in the current form - it
does a recursive search of all parents of a temp directory, but
we have no control at all on the contents of directories outside
of the ones we control.

So, I tweaked code to
- allow us to optionally truncate the search at some path
- make two levels of directories in the tests
- stop search after checking the outer level

Reviewed By: grievejia

Differential Revision: D31800614

fbshipit-source-id: 6e20e8074699f9ac13ea7e8ad3a0dc3b093d8877