Fix unit tests that are broken on OSX due to links
In OSX 11.6, it looks like the output of `tempfile.TemporaryDirectory()`
is not necessarily a true, fully-resolved path but the output of
`Path.cwd()` is. As a result, it's necessary for us to fully-resolve
the temporary directory in a couple of the unit tests to avoid
mismatched paths.

```
./scripts/run-python-tests.sh
```
which reported two failures prior to this change, none after.