Add external query feature for SQL queries
Summary:
This feature is useful to identify flows which goes through queries and for this reason the output of the taint may not be related anymore to the actual user input.
Used the TaintInTaintOut notation instead of AttachToTito to make explicit the TITO flow through these parameter without relaying on the inferred flow.

Reviewed By: gbleaney

Differential Revision: D31148095

fbshipit-source-id: 0fc147a8d19edaeaaa3da40e4e06841d3eaf40b4