Fix unit tests that are broken on OSX due to links (#517)
Summary:
In OSX 11.6, it looks like the output of `tempfile.TemporaryDirectory()`
is not necessarily a true, fully-resolved path but the output of
`Path.cwd()` is. As a result, it's necessary for us to fully-resolve
the temporary directory in a couple of the unit tests to avoid
mismatched paths.

Pull Request resolved: https://github.com/facebook/pyre-check/pull/517

Test Plan:
```
./scripts/run-python-tests.sh
```
which reported two failures prior to this change, none after.

Reviewed By: dkgi

Differential Revision: D31896292

Pulled By: stroxler

fbshipit-source-id: ef43716d12a553304f23666ae1905d3bdaaa3d16