Infer: remove interprocedural logic from client
Summary:
We no longer think we'll ever use interprocedural mode for
pyre infer (we may use an incremental-like AST rewrite
eventually, but we probably wouldn't call it "interprocedural").

So, cleaning up the code. I got rid of most of the ocaml logic
in the previous commit, here I remove the flag from the client
and from the ocaml config serialization raw data.

NOTE: this won't be safe to land until D31731942 (https://github.com/facebook/pyre-check/commit/029bbd8895bdad07abbf1ace362ce48014da7fa5) is landed and
released to fbcode, because the current binary has mode as a non-optional
entry in the json. Tests will catch problems, so we can just rebase when we
have an upgrade finished and if tests pass we're good to go!

Reviewed By: grievejia

Differential Revision: D31733049

fbshipit-source-id: 7724f8c87a1f1747afe84c00f5964e17b1df954a