Differentiate btw. callable and module not in env (#511)
Summary:
Lately, we did not differentiate between a module not being in the
environment and functions not being in the environment and displayed the
same error message - "`this` is not part of the environment".

Adds logic to differentiate between callable not being present and
module not being present and display a different error if the module is
present but not the callable. This happens when a user writes a model
file with an invalid function/callable name.

Pull Request resolved: https://github.com/facebook/pyre-check/pull/511

Test Plan:
- compile the source from scratch following instructions in pyre installation docs
- write up a model with or add to a model an invalid callable name: eg) add the following non existent callable model to any model file:
```python
def flask.this_is_a_not_present_function(a: TaintSink[SQL])
```
- point your `.pyre_configuration` to the path containing such models
- run `pyre analyze` (where pyre is an alias to `python3 -m client.pyre`, ie, the module in the source)

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>
Fixes: https://github.com/MLH-Fellowship/pyre-check/issues/81

Reviewed By: onionymous

Differential Revision: D31654047

Pulled By: 0xedward

fbshipit-source-id: 9f7aa483df3e8e1e04d3659aa5cdbe7cd96f450b