Implement sanitize taint transforms as sets
Summary:
# Problem
Sanitize taint transforms can easily blow up due to combinatory explosion.
For instance, `Not[A]:Not[B]:C` and `Not[B]:Not[A]:C` are two distinct taint kinds.

# Current solution
To prevent this, we are sorting that list at some key points of the algorithm, but it might not be enough.

# Proposed solution
To circumvent this definitely, lets implemented sanitized taint transforms as sets instead of lists.

Reviewed By: dkgi

Differential Revision: D31936500

fbshipit-source-id: 7dc3afe1cd26c1486f5fb05908d5a4504881d476