References: mapping to simplify qualification if possible
Summary:
Idea is to only keep qualifiers where they are necessary to disambiguate.

For a simple example, if we have two references "Foo.a" and "Foo.b", the utility function will return a mapping "Foo.a" -> "a" and "Foo.b" -> "b".

On the other hand, if we have two references "Foo.a" and "Bar.a", we need to keep the qualification.

See tests for more examples.

This is intended for use to simplify our error messages. That itself will be a somewhat trivial but tedious change because most of our tests rely on the string output. Will do that once we are settled on the disambiguation logic itself.

Reviewed By: grievejia

Differential Revision: D31941471

fbshipit-source-id: c3c9fc853b49568174ff9871ba5ff833f6aa8bb0