Vendor ApplyTypeAnnotationsVisitor from LibCST.
Summary:
This commit vendors the current `ApplyTypeAnnotationsVisitor`
logic from the `main` branch of `Instagram/LibCST` into
`pyre/client`.

## Overall rationale for vendoring

Why do we need this? Because LibCST has a release cycle that
isn't instantaneous, and we are actively developing changes in LibCST's
ApplyTypeAnnotationsVisitor specifically to help pyre infer.

The idea is that we'll be able to write and test changes to the
code in the Instagram code, and mirror the changes into pyre
to get benefits immediately.

We include a header on the file explaining where it comes from,
including the commit of LibCST.

## What's the immediate benefit of vendoring right now?

By vendoring today, we'll get recent changes
that cause LibCST to leave existing return annotations whenever
the output of the backend is missing. The current release would nuke
existing return annotations, which forced us to include both inferred and
given annotations in the output of the backend.

I'm intending to stack commits that will take advantage of this by changing
`pyre infer` to skip given annotations, which will reduce the frequency with
which incorrect import handling affects us.

## What about LibCST tests?

I'm not vendoring the LibCST tests because we will only ever commit
vendored code that is already passing LibCST tests. But we could revisit
this if we discover there's some benefit to duplicating the tests as well.

Reviewed By: grievejia

Differential Revision: D31961532

fbshipit-source-id: 5d77e0c64b1aea8b57fd8bb286287990cdf9b617