Tweak handling of future annotations flag
Summary:
At the time when we introduced the --no-future-annotations
flag, we expected to turn it on by default but this has not
proved fruitful thus far, for several reasons:
- because we so often accidentally produce diffs with
  no actual annotations, we had to turn it off in prod to
  avoid spamming users with diffs that just added this one
  line.
- in order to fix problems with import handling in LibCST,
  I have an open PR on LibCST that would break the logic
  anyway.

Also, at the time I wrote the code I had forgotten how to
make a boolean flag have both explicit-True and explicit-False
options (this problem is pervasive in pyre, in my opinion it's
a bit of a click antipattern; it's a bigger problem for infer
because so many flags control experimental features and we
aren't sure the best defaults yet).

I think we'll still want this option in the future so I'd
rather not remove it altogether just yet, but it would be
nice to make the default `use_future_annotations=False`.
So, what I'm doing is
- naming the cli argument the same as the python variable
- allowing both explicit True and explicit False
- making the default False, which is what we already do
  in all prod use cases.

Reviewed By: grievejia

Differential Revision: D31975203

fbshipit-source-id: ba75ef5458fa05cb1b6040b0da25df744648eef2