Only report inferred types, not given
Summary:
Because LibCST still has imperfect logic for translating
the types in stub files to the imports it needs to generate
in code, it's much better if we can minimize the number of
types we have to deal with when performing codemods.

So, in this diff I'm modifying the ocaml logic to output
`null` for any annotation that was provided for us by the
existing python code (as opposed to one for which we actually
performed type inference).

# Discussion: this does change stub output. Do we care?

This diff should be a strict improvement in the quality of
codemods that `pyre infer` produces.

That said, it is a breaking change for anyone wanting to kick of
the process of making a stub file using `pyre infer`, because now
the stubs would be generated without existing annotations.

I don't consider this a valid concern because:
- `pyre infer` has *always* completely omitted any functions that
  didn't get new annotations, which means it was never really
  useful for bootstrapping true stub files.
- `pyre infer` has always spit out "stub files" that use fully
  qualified names without import, which isn't legal python, so again
  it's never really been useful for creating stub files.

But if this does become a concern, we could always come back to the
code and functorize the `LocalResult` / `GlobalResult` types so that
the serialization of `TypeAnnotation.t` can either include or exclude
given annotations.

Reviewed By: grievejia

Differential Revision: D31968823

fbshipit-source-id: 271f00a3f4ca6a7e5aec8998c86272561992bf42