Don't complain when using `pyre_extensions.Generic`
Summary:
Problem: When we use `class Foo(pyre_extensions.Generic[T])`, Pyre says that `pyre_extensions.generic.Generic` is not subscriptable.

This is because we were translating `pyre_extensions.Generic` to `typing.Generic`, but not `pyre_extensions.generic.Generic`.

The change to `queryTest.ml` is harmless. There is no unintended runtime behavior change.

Reviewed By: stroxler

Differential Revision: D31991682

fbshipit-source-id: a593703818ff48a5276e7e6903796359922a24b5