Avoid metaclass in `pyre_extensions.Generic`.
Summary:
Problem: Our current implementation of `Generic` uses a metaclass. This causes problems when a child class is generic and a base class has a custom metaclass.

```
metaclass conflict: the metaclass of a derived class must be a (non-strict) subclass of the metaclasses of all its bases
```

Ran into this when the base class was abstract, which means it has metaclass `ABCMeta`.

The way `typing.Generic` gets around this is by using `__class_getitem__`, available from 3.7 onwards. So, do the same.

Reviewed By: stroxler

Differential Revision: D31991683

fbshipit-source-id: b21f436d9844d9c0e32f5b2a61137e04e7cf0871