Change `annotation_matches_constraint` to look at the AST rather than resolved type
Summary:
Previously we were using `Resolution.resolve_reference` to get the type of attributes and using that for the `type_annotation` clause match. The problem is that when Pyre parses `Annotated` types it discards the entire annotation after the first argument, i.e. it treats `Annotated[x, A, B]` as just `Annotated[x]`, so we cannot match on `A` or `B`.

Instead, let's look at the AST and the parsed expression form of the annotation. It's a little dodgy but we can match on the annotation part of `Annotated` types without having to change how Pyre understands the type, which is probably for the best.

Note that the annotations _will not_ show up in `via-type-of`. e.g. in the integration test case:
For something like:
```
class Test1_C:
    a: Annotated[Optional[float], Color.RED] = None
    b: Annotated[Optional[float], Color.BLUE] = None
```

With `ViaTypeOf` the type of `Test1_C.a` still shows up as
```
"always-via-type": "typing.Annotated[typing.Optional[float]]"
```

but we can at least find them using the DSL with:
```
type_annotation.matches(".*Annotated\[.*Color\..*")
```

Reviewed By: arthaud

Differential Revision: D31986451

fbshipit-source-id: 35e10c6a91e5b945a4ffde60a256626f7aa93261