Clean up unclassified errors
Clean up unclassified errors by mitigating each errors to corresponding
error types instead of a single unclassified error type.

Changes are done in such a way that the messages are made identical to
previous ones as much as possible. Thus, this is just a implementation
change which should have no effect on the output.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>