Remove taint that can never match any rule
Summary:
In previous patches, we implemented sanitize taint transforms, which add information in the taint kind about what sinks are not valid for a given source, or what sources are not valid for a given sink. For instance, `Not[UserControlled]:RemoteCodeExecution`.

In this patch, we automatically discard taint that can never match a rule. For instance, `Not[UserControlled]:RemoteCodeExecution` can be discarded if there are no other rules than `UserControlled -> RemoteCodeExecution`.

Reviewed By: dkgi

Differential Revision: D32031178

fbshipit-source-id: fbc9cd6c3b96ca1cee9333cd94c2b8a694acac9b