Make it possible to load a project in utop
Summary:
This is a first stab at getting to the point where it's feasible
to use utop as an interactive environment for looking at a
pyre project and interactively understanding the APIs used to
build and query the environment.

It's very helpful, for these purposes, to be able to
interactively load the arguments from a `pyre -n check` command's
json output and then be able to extract a `Configuration.Analysis.t`;
this allows us to easily recreate in utop the same configuration
that powered a `pyre check` command.

Reviewed By: dkgi

Differential Revision: D32069574

fbshipit-source-id: 54cc16a5a95eb2b522c9513c50b7cf67d7a00cb2