Infer: use LibCST for --use-future-annotations flag
Summary:
After bumping to the latest LibCST version, we no longer
want to handle `from __future__ import annotations` by inserting
it into the stub file - the new LibCST will ignore dangling
imports in the stub anyway, and instead has a flag to control this
behavior.

An added benefit is that we now have simpler code:
- Conceptually, we now *never* add any imports to the stub, we
  exclusively use fully-qualified names. That's easier to think about
- In terms of lines of code, we were able to remove a couple methods
  from the ModuleAnnotations class :)

Since the `use_future_annotations` flag is part of
`StubAnnotationOptions`, I needed to start passing `options`
to the code generation task. At this point it made sense
to also move the `debug_infer` flag into the options, since the
whole point of having an options dataclass is to avoid passing
a lot of boolean flags around individually (which makes the
code harder to skim and and harder to change).

Reviewed By: dkgi

Differential Revision: D32054516

fbshipit-source-id: 3306d7257cc39395e153a2a36d27d61dcb279efd