Adds key validation to taint.config files
Adds validation of required and optional keys of taint.config files to
taintConfiguration parser in the server.

Previosuly, we didn't do such validation and ignored invalid keys in the
file and errored out with "not so pretty" error message when a required
key was absent.

Now we error out with a pretty error message and show a warning about
invalid keys while continuing through the parsing process.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>