Accept `hover` LSP requests.
Summary:
Problem: Pyre tells the LSP server that it accepts `hover` requests. So, we need to respond to them.

This diff parses the request and returns a dummy response.

In case we can't find any type, we return an empty hover response. The IDE will basically not show anything for us in the hover card.

Next diff: Populate types using `pyre query` and look up the type on each hover.

Reviewed By: grievejia

Differential Revision: D31735965

fbshipit-source-id: babe9bcbf15232f7414052e38c493b504b80f454