Add path to be queried upon saving a file.
Summary:
Problem: We want to populate the type-lookup data structure.

When a file is updated (saved, opened, etc.), `PyreServer` will enqueue it into a shared queue.

A background pyre query manager will be responsible for waiting on the queue and querying for the types in those files.

See base diff for more context.

Reviewed By: grievejia

Differential Revision: D31735964

fbshipit-source-id: 6167d3dd3c07ff649e2c953c49af475e0f5ffd90