Run a pyre-query-caller in the background.
Summary:
Problem: The previous diffs allow us to respond to a hover request by looking up the type for that file position. Now, we need to populate the type-lookup data structure.

Design choices:
+ Edit: I'm going for a fresh server connection for every query. Having a persistent connection wasn't actually much faster and was a lot less clean. This means the query handler can be independent of the server handler.

+ I'm opening a different connection because the existing one is being used to subscribe to type error diagnostics.

Note that we do this in a pull-driven manner so that we can only query for the currently-active file when it saved or opened. This is better than the push-driven approach which will inundate the persistent client with type-info for all open files on every save (and will need the server to keep track of the open files, which defeats the purpose of the subscription).

Remaining:
+ In the next diff, I'm only running the handler when hover is enabled.
+ Restart the query handler when restarting the server.

Reviewed By: grievejia

Differential Revision: D31735959

fbshipit-source-id: d97741090b9eb976333a6bf9ba1b606b5dcd92f1