Query pyre for types in files.
Summary:
Problem: We need to respond to the LSP `hover` request with the type at that location.

The previous diff added the interval tree class. This diff adds a function to populate the interval tree using a `Query` request to the server.

This function will basically be called by the background task whenever there is a path to be queried for.

Reviewed By: grievejia

Differential Revision: D31735966

fbshipit-source-id: d31320f036112316ebc8337bfbb2a0c02e61f5fe