Use the same interface for apply_call_targets in the forward and backward analysis
Summary:
In the forward analysis, `apply_call_targets` takes a callee and arguments, but in the backward analysis, it takes a call expression.
Let's make it consistent by taking a callee in the backward analysis.
Note that we fabricate a call expression anyway to resolve the type, but that expression is purely symbolic and might not make sense in the actual code, especially when handling constructors and property setters.

Reviewed By: dkgi

Differential Revision: D32225059

fbshipit-source-id: 1fc80a00e05817cac8b5d22a23bc25ed579cdb58