Avoid mangling class decorators and bases with private-like naming
Summary:
We use an expression transformer to go through all private names in a class and mangle them in accordance to python runtime behavior.

However, re-using the transformer logic only gives us a few knobs to control, like yes/no binary decision on whether or not to transform all the children of a statement. Expressions within class bases and decorators are considered children of the class statement, but we really only want to be transitively transforming all expressions in the class body.

To avoid mangling class bases (which was the source of a user-reported bug where the class hierarchy did not behave as expected due to a class named with a leading double underscore), I believe the simplest adjustment we can make is to save the class metadata before the transformer goes in to mangle all nested expressions, and re-set the decorator and bases when we are popping that class off the stack of mangling prefixes.

We avoid storing the body because it's not needed, and is a lot more data than just the class metadata.

Reviewed By: grievejia

Differential Revision: D32139893

fbshipit-source-id: 5dd81763b8b92b8619f588ad69acbe4a5225c540