Extract `apply_call_target` from `apply_call_targets`
Summary:
# Problem
When handling constructor calls, we re-analyze the arguments twice because the simulate a call to `__new__` followed by `__init__`.

# Solution
We should analyze arguments only once and then properly forward the taint when handling `__new__` and `__init__` separately. For that, let's extract the `apply_call_target` function out of `apply_call_targets` first.

Reviewed By: dkgi

Differential Revision: D32226957

fbshipit-source-id: b911473a7fc7a06771d3d4db8d012af2b27cb519