Add tests of the generator cases we should handle
Summary:
I don't think it's typical for us to have a "known_failures"
test function, but as a temporary tool for guiding development
I find it helpful, because it lets me see at a glance not
only which cases we can't type correctly but what the actual
output of type checking is for those cases.

My plan is to begin patching the incorrect results and moving
them out of this "known errors" test and into the preexisting
tests until there are no known failures left and we can remove
this test.

Reviewed By: dkgi

Differential Revision: D32268844

fbshipit-source-id: 6f3184072274cb96a93b24503592aad46c23ec0d