Minimize mutual recursion in `typeCheck.ml`
Summary:
There was no real reason for the `initial` function in
the `typeCheck.ml` `State` module to be recursive - it
was only that way because it called into `forward_expression`,
but this is easily fixed by moving it to the bottom.

I like the new code better because
- the order is clearer: the `State` interface
  functions `initial`, `forward`, and `backward` are next
  to one another
- the structure is clearer: none of the public `State` functions
  involves recursion, only the inner helper functions that traverse
  expressions and statement.

This commit is a pure reorder plus changing of where the recursive
block happens; no logic changes whatsoever.

Reviewed By: grievejia

Differential Revision: D32265182

fbshipit-source-id: d2a3eeebef9d7649403cde1314f2d2736b5121a2