Use iterable logic for YieldFrom handling
Summary:
For reasons I don't understand, we were manually constructing
a fake call to `__iter__` and then checking the resulting type
against `typing.Iterator` in `yield from` handling, both in
`typeCheck.ml` and `typeInferenceLocal.ml`.

This is verbose and a bit strange: anything with an `__iter__`
method returning an `Iterator` is an `Iterable`, so why not
just check that directly?

In general I'm pretty sure we should *never* special-case
`Iterator` in pyre: the Iterable-vs-Iterator distinction is
an implementation detail that allows user code to separate
out the initial object from the iterator. Everything we need
to know about that is encoded in `typing.pyi` using normal
python rules. Syntactically, we should only ever care about
`Iterable`, since non-Iterator-Iterable values are permitted
in every single python form.

Reviewed By: grievejia

Differential Revision: D32265018

fbshipit-source-id: 010c7a8e6c9114b94646672a1f8c7775673b14fd