Remove dead function async_generator_value
Summary:
I'm not sure why this function ever existed, it doesn't
really make sense to turn an async generator into a regular
generator (as far as I know the two types are never compatible)
and it's definitely unused in the current code.

Having dead functions to maintain is going to make it harder to
fix our generator handling code. Let's kill it!

Reviewed By: arthaud

Differential Revision: D32250509

fbshipit-source-id: 099cbcbcaf917c40c67a29886f0d1a5f505c0c75