Extract iterable value type calculation
Summary:
We extract the value extracted from an iterable in many
places in pyre, because multiple built-in language
constructs (for loops, generator expressions, yield from
expressions, etc) all rely on this and our code has to
implement every one of them.

This is cool, nothing wrong with having it in many places.
But instead of using a well-named and concise function call,
we're currently hardcoding the extraction out of a very low
level `extract_type_parameters`.

By pulling this out into a helper function, we
- make the code simpler, both by line count and by
  the level of nested match statements
- make it easier to find all the iterable special cases
  with a code search

Reviewed By: grievejia

Differential Revision: D32265183

fbshipit-source-id: 9cb9ff77e5f5ad362fdd0860665319a4d49428b3