Extract `apply_call_targets_with_arguments_taint` from `apply_call_targets`
Summary:
# Problem
When handling constructor calls, we re-analyze the arguments twice because we simulate a call to `__new__` followed by `__init__`.

# Solution
We should analyze arguments only once and then properly forward the taint when handling `__new__` and `__init__` separately.
For that, we need to refactor `apply_call_targets` to call some parts of it.
This diff extract `apply_call_targets_with_arguments_taint` from `apply_call_targets` in the forward analysis, which takes a list of taint for each argument and applies the given list of call targets.
We also extract `apply_call_targets_and_return_arguments_taint` from `apply_call_targets` in the backward analysis, which takes the call taint and return a list of taint for each argument.

Reviewed By: dkgi

Differential Revision: D32267707

fbshipit-source-id: 13c229f0736eaa55e6f735106da5b3d2484a8db7