Analyze arguments of constructor calls only once
Summary:
This diff fixes a performance issue due to analyzing the arguments of a constructor calls twice, once for the implicit call to `__new__` and once for the call to `__init__`.
This diff also fixes a bug where the call to `__init__` does not actually receive the taint created from `__new__`.

Reviewed By: dkgi

Differential Revision: D32271844

fbshipit-source-id: 07c313922f7f695cec9fa8147c6fddce54e50db1