Generate a model for the `__new__` constructor of named tuples
Summary: We generate a precise model for the `__init__` constructor of named tuples, which properly propagates the taint of arguments to their appropriate attribute. However, we do NOT generate a model for `__new__` and hence get an obscure model for it, which would propagate all taint to the whole object (leading to collapsing). This is not a problem right now because our handling of constructor is buggy, but it will show up in the follow-up diff. Let's fix it by generating an empty model for `__new__`.

Reviewed By: dkgi

Differential Revision: D32271680

fbshipit-source-id: a9d54700b338b9b351657248abab3cb98c3f7718