Allow testing of qualify_statement directly (2/n)
Summary:
# Context
See top of the stack: D32221881.

# This diff
Make qualify_{statement,expression,...} etc top-level.

Put `relative` and `qualifier` which are needed from SourcePath in a module Context to Qualify. Note that since these don't need to modified during the qualification pass, it doesn't make sense to pass it around (adding it in the scope, or otherwise).

Reviewed By: grievejia

Differential Revision: D32225287

fbshipit-source-id: 290a9859055482bbdcf716922727cfa6a06a940f