Allow testing of qualify_statement direclty (1/n)
Summary:
# Stack

This stack makes it easier to write tests for `qualify` at lower levels of AST, as creating a Source.t object directly is very tedious and verbose.

Currently, we do not allow for a way to test qualification of anything other than a `Source.t`. The existing tests are written assuming the availability of a parser, which makes this bearable. This is currently not an option for anything that requires the new parser (see D32194094 (https://github.com/facebook/pyre-check/commit/2e324152571675c48e0238e651351f80e91b57ee)). Even in the absence of this requirement, it is perhaps better from the point of view of unit testing to not need this dependency.

It requires some, but not too much, refactoring on the `preprocessing.ml` side.

One down side is that it will require Preprocessing module expose some functionality only for testing purposes. We limit the impact as much as possible, but this is unforunately unavoidable.

# This diff

Put things inside a Qualify module, so when we expose it externally; things like the record types and individual functions will not leak into the Preprocessing namespace directly.

Reviewed By: grievejia

Differential Revision: D32221881

fbshipit-source-id: 8b9d1e43ec575bf2405f7a5bab27952ed329f0d6