yield exprs [1/n] type_of_generator_return_value
Summary:
Like the previous commit adding type_of_generator_send_value,
this will help us properly resolve expression types - in this case,
the type to which a `yield from` expression will resolve.

According to PEP 380, such an expression may be used against not
only generators but also any iterable type. It resolves to the
return type of a generator, otherwise to `None`.

Reviewed By: grievejia

Differential Revision: D32288479

fbshipit-source-id: 7ac719a1c32230d7856f44c9a558e86a53b516b5