yield exprs [0/n] type_of_generator_send_value
Summary:
Add a function that, given a return annotation on a
generator function (to be clear, when we say a "generator function"
this is determined syntactically based on existence of a yield
or yield from, not based on a type annotation), will give us the
appropriate send type (i.e. the type to which a `yield` expression
should evaluate).

I fall back from Generator to AsyncGenerator to NoneType because
that seems likely to give the most usable results for type checking
purposes: whenever someone uses a non-generator annotation on
a generator function, that implies the send type must be None

Reviewed By: grievejia

Differential Revision: D32286095

fbshipit-source-id: 40d7bd25181da9749956cdce1e4a22cd28ab4167