yield exprs [3/n] fix the resolution of yield expressions
Summary:
This diff completes the change to correctly evaluate the
type of `yield` and `yield from` expressions:
- according to PEP 342 (which added send to generators)
  a `yield` expression evaluates to the send type of
  this function if annotated as a generator, else `None`
  (e.g. if a generator function is annotated as an iterator)
- according to PEP 380 (which introduced `yield from`,
  a `yield from` expression evaluates to the return type of
  the underlying generator if the underlying expression
  is a generator, else `None` (e.g. if the underlying expression
  is some other iterable type)

This closes the expression-type gaps that were recorded in
`check_generator_known_failures`.

Note that we are still not correctly handling several other
things, such as validating the send and yield types for `yield from`
expressions.

Reviewed By: grievejia

Differential Revision: D32293164

fbshipit-source-id: d9fb3afd79ffc04d70c119a360968175dd037eb7