yield validation [0/n] Extract validate_return function
Summary:
Now that we've got yield and yield from expressions evaluating
to the correct types, the next step is to ensure that we correctly
handle the side conditions:
- any yield argument needs to be compatible with the yield type.
  We currently partially handle this (we assume the generator
  annotation has to be compatible with Generator[yeilded, None, None]
  and only validate bare yield statement), but we need to
  fully handle it which means
  - we need to *not* impose restrictions on the send or return type
  - we need to validate all yield expressions, not just bare ones
- both the yield type and send type of any generator that we
  integrate with using a `yield from` need to be (covaraiantly
  and contravariantly, respectively) compatible with the current
  generator function's type. The return types are not related.

A prerequisite to handling all of this is to push yield expression
vs return validation logic into `forward_expression`, which requires
that `validate_return` be a top-level (private) function in the
`State` module.

This commit is pure refactor, no logic changes whatsoever. A bonus
is that now `forward_statement` is an almost-reasonable 300 lines,
which makes it easier to skim the code in my opinion.

Reviewed By: grievejia

Differential Revision: D32299540

fbshipit-source-id: de43da487844f37a4cc559fbe53ed0cba0e9c2cb